TlsService

+++ RS.Schannel.Manager.API/NativeMethods/SslProviderProtocolId.cs
@@ -10,7 +10,7 @@
 #pragma warning disable CS1591,CS1573,CS0465,CS0649,CS8019,CS1570,CS1584,CS1658,CS0436,CS8981
 namespace Windows.Win32;
 
-public enum SslProviderProtocolId : uint
+public enum SslProviderProtocolId : ushort
 {
     SSL2_PROTOCOL_VERSION = 0x0002,
     SSL3_PROTOCOL_VERSION = 0x0300,
